services/embedding-service/requirements.txt
kafka-python
chromadb
sentence-transformers
tqdm
sqlalchemy
psycopg2-binary
fastapi
uvicorn